Connect contracts crate to the MIRAI static analyzer.
This adds a cargo feature mirai_assertions to the contracts crate. If this feature is on,
contracts will emit MIRAI macros instead of the standard assert! style macros.

To execute integration tests with MIRAI macros, use

    cargo test --tests --features=mirai_assertions

The tests will use mocks of the MIRAI macros which map back to regular
runtime assertions.